+++ packages/cozeloop-ai/src/tracer/constants/index.ts
@@ -1,24 +1,19 @@
 // Copyright (c) 2025 Bytedance Ltd. and/or its affiliates
-
-import { createContextKey } from '@opentelemetry/api';
-
 // SPDX-License-Identifier: MIT
+
 export const ROOT_SPAN_PARENT_ID = '0';
 
 export const COZELOOP_LOGGER_TRACER_TAG = 'Tracer';
 
-export const COZELOOP_TRACE_SPAN_USER_ID_KEY = createContextKey(
-  'cozeloop.trace.span.user_id',
-);
-export const COZELOOP_TRACE_SPAN_MESSAGE_ID_KEY = createContextKey(
-  'cozeloop.trace.span.message_id',
-);
-export const COZELOOP_TRACE_SPAN_THREAD_ID_KEY = createContextKey(
-  'cozeloop.trace.span.thread_id',
-);
-
 export enum COZELOOP_TRACE_IDENTIFIER {
   LOOP = 'cozeloop-js',
+}
+
+export enum COZELOOP_TRACE_PROPAGATION_HEADERS {
+  W3C_TRACEPARENT = 'traceparent',
+  W3C_TRACESTATE = 'tracestate',
+  COZELOOP_TRACEPARENT = 'X-Cozeloop-Traceparent',
+  COZELOOP_TRACESTATE = 'X-Cozeloop-Tracestate',
 }
 
 export enum COZELOOP_TRACE_SPAN_STATUS_CODE {
@@ -34,6 +29,8 @@ export enum COZELOOP_TRACE_BASIC_TAGS {
   PARENT_SPAN_ID = 'cozeloop_parent_span_id',
   SPAN_WORKSPACE_ID = 'cozeloop_workspace_id',
   SPAN_ULTRA_LARGE_REPORT = 'cozeloop_ultra_large_report',
+  SPAN_CUSTOM_ROOT_SPAN_ID = 'cozeloop_custom_root_span_id',
+  SPAN_RUNTIME_SCENE = 'cozeloop_span_runtime_scene',
 
   // Reserved tags
   SPAN_LOG_ID = 'cozeloop_span_log_id',
@@ -44,9 +41,7 @@ export enum COZELOOP_TRACE_BASIC_TAGS {
 
 export enum COZELOOP_TRACE_BUSINESS_TAGS {
   // Common tags
-  USER_ID = 'user_id',
-  MESSAGE_ID = 'message_id',
-  THREAD_ID = 'thread_id',
+  ERROR_MESSAGE = 'error',
 
   // Tags for Model span
   /** The timestamp of the model's first response when using stream response. The unit is microseconds. */